refactor collapsible table summary
